Update substitutions, add options, and modify steps
Reformatted and updated the `substitutions` section with specific values for `_SERVICE_NAME`, `_REGION`, `_PROJECT_ID`, and `_REPO_NAME`. Added `options` section with `logging: CLOUD_LOGGING_ONLY` to enable Cloud Logging. Modified Docker authentication step to use `docker login` with OAuth2 access token instead of `gcloud auth configure-docker`. Added a new step to tag and push a release tag to GitHub, configuring Git, creating a tag with the current date and time, and pushing the tag to the origin repository. Removed the redundant `options` section at the end of the file.