Use $BUILD_ID for Docker image tagging in cloudbuild.yaml
Modified the Docker image build and push steps in the cloudbuild.yaml
file to use the $BUILD_ID environment variable for tagging the Docker
image instead of a timestamp-based tag. This change ensures consistent
and unique tagging of Docker images, aiding in better tracking and
management of images associated with specific builds.